requirements.txt
torch>=1.10
torchvision
pandas
matplotlib
Pillow
